Moved encoding code into a separate package Added WmmEncoder (unfinished)
